analog: cast integer divide to type int for nbfm_tx

+++ gr-analog/python/analog/nbfm_tx.py
@@ -52,7 +52,7 @@ class nbfm_tx(gr.hier_block2):
         do_interp = audio_rate != quad_rate
 
         if do_interp:
-            interp_factor = quad_rate / audio_rate
+            interp_factor = int(quad_rate / audio_rate)           # force integer
             interp_taps = filter.optfir.low_pass(interp_factor,   # gain
                                                  quad_rate,       # Fs
                                                  4500,            # passband cutoff
